update comment on delete
