+++ app/src/main/java/org/stepic/droid/view/fragments/CommentsFragment.kt
@@ -1,13 +1,10 @@
 package org.stepic.droid.view.fragments
 
-import android.app.Dialog
 import android.os.Bundle
 import android.support.design.widget.CoordinatorLayout
 import android.support.design.widget.FloatingActionButton
-import android.support.v4.app.DialogFragment
 import android.support.v4.app.Fragment
 import android.support.v4.widget.SwipeRefreshLayout
-import android.support.v7.app.AlertDialog
 import android.support.v7.app.AppCompatActivity
 import android.support.v7.widget.LinearLayoutManager
 import android.support.v7.widget.RecyclerView
@@ -17,7 +14,6 @@ import android.text.style.ForegroundColorSpan
 import android.view.*
 import android.widget.ProgressBar
 import android.widget.Toast
-import com.squareup.otto.Bus
 import com.squareup.otto.Subscribe
 import com.yandex.metrica.YandexMetrica
 import org.stepic.droid.R
@@ -32,11 +28,9 @@ import org.stepic.droid.util.AppConstants
 import org.stepic.droid.util.ColorUtil
 import org.stepic.droid.util.ProgressHelper
 import org.stepic.droid.view.adapters.CommentsAdapter
-import org.stepic.droid.view.custom.LoadingProgressDialog
+import org.stepic.droid.view.dialogs.DeleteCommentDialogFragment
 import org.stepic.droid.view.util.ContextMenuRecyclerView
-import org.stepic.droid.web.CommentsResponse
 import org.stepic.droid.web.DiscussionProxyResponse
-import org.stepic.droid.web.IApi
 import org.stepic.droid.web.VoteResponse
 import retrofit.Callback
 import retrofit.Response
@@ -381,11 +375,15 @@ class CommentsFragment : FragmentBase(), SwipeRefreshLayout.OnRefreshListener {
         if (!commentManager.isEmpty()) {
             showEmptyState(false)
         }
-        val needInsertLocal = needInsertOtUpdateLate
+        val needInsertLocal : Comment?= needInsertOtUpdateLate
         if (needInsertLocal != null && (!commentManager.isCommentCached(needInsertLocal.id) || (needInsertLocal.parent != null && !commentManager.isCommentCached(needInsertLocal.parent)))) {
             val longArr = listOf(needInsertLocal.id, needInsertLocal.parent).filterNotNull().toLongArray()
             commentManager.loadCommentsByIds(longArr)
         } else {
+            //we have only our comment.
+            if (needInsertLocal != null) {
+                commentManager.updateOnlyCommentsIfCachedSilent(listOf(needInsertLocal))
+            }
             needInsertOtUpdateLate = null
             commentAdapter.notifyDataSetChanged()
         }
@@ -462,62 +460,6 @@ class CommentsFragment : FragmentBase(), SwipeRefreshLayout.OnRefreshListener {
     @Subscribe
     fun onFailDeleteComment(event: FailDeleteCommentEvent) {
         Toast.makeText(context, R.string.fail_delete_comment, Toast.LENGTH_SHORT).show()
-    }
-
-    class DeleteCommentDialogFragment : DialogFragment() {
-
-        @Inject
-        lateinit var api: IApi;
-
-        @Inject
-        lateinit var bus: Bus;
-
-        companion object {
-            private val COMMENT_ID_KEY = "comment_id_key"
-            fun newInstance(commentId: Long): DialogFragment {
-
-                val args = Bundle()
-                args.putLong(COMMENT_ID_KEY, commentId)
-                val fragment = DeleteCommentDialogFragment()
-                fragment.arguments = args
-                return fragment
-            }
-        }
-
-        override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-            MainApplication.component().inject(this)
-            val loadingProgressDialog = LoadingProgressDialog(context)
-            val commentId = arguments.getLong(COMMENT_ID_KEY)
-            val builder = AlertDialog.Builder(activity)
-            builder
-                    .setTitle(R.string.title_confirmation)
-                    .setMessage(R.string.delete_comment_detail)
-                    .setPositiveButton(R.string.delete_label) { dialog, which ->
-                        ProgressHelper.activate(loadingProgressDialog)
-                        YandexMetrica.reportEvent(AppConstants.DELETE_COMMENT_CONFIRMATION)
-                        api.deleteComment(commentId).enqueue(object : Callback<CommentsResponse> {
-                            override fun onResponse(response: Response<CommentsResponse>?, retrofit: Retrofit?) {
-                                ProgressHelper.dismiss(loadingProgressDialog)
-                                if (response?.isSuccess ?: false) {
-                                    val comment = response?.body()?.comments?.firstOrNull()
-                                    comment?.let {
-                                        bus.post(NewCommentWasAddedOrUpdateEvent(it.target!!, it))
-                                    }
-                                } else {
-                                    bus.post(FailDeleteCommentEvent())
-                                }
-                            }
-
-                            override fun onFailure(t: Throwable?) {
-                                ProgressHelper.dismiss(loadingProgressDialog)
-                                bus.post(InternetConnectionProblemInCommentsEvent(null))
-                            }
-
-                        })
-                    }.setNegativeButton(R.string.cancel, null)
-
-            return builder.create()
-        }
     }
 
 }
